+++ actionmailer/lib/action_mailer/message_delivery.rb
@@ -85,26 +85,6 @@ module ActionMailer
       message.deliver
     end
 
-    def deliver! #:nodoc:
-      ActiveSupport::Deprecation.warn(<<-MSG.squish)
-        `#deliver!` is deprecated and will be removed in Rails 5. Use
-        `#deliver_now!` to deliver immediately or `#deliver_later!` to
-        deliver through Active Job.
-      MSG
-
-      deliver_now!
-    end
-
-    def deliver #:nodoc:
-      ActiveSupport::Deprecation.warn(<<-MSG.squish)
-        `#deliver` is deprecated and will be removed in Rails 5. Use
-        `#deliver_now` to deliver immediately or `#deliver_later` to
-        deliver through Active Job.
-      MSG
-
-      deliver_now
-    end
-
     private
 
       def enqueue_delivery(delivery_method, options={})